pyproject.toml
[project]
name = "langgraphcourse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "python-dotenv>=1.1.1",
]
